allow custom resolv.conf handling via hooks
The primary goal of this is to allow for no resolv.conf handling at
all. This is useful on a host that functions as a router for clients
that need VPN access, in which case the router should not actually
have its own DNS tampered with.

In this case, the user will need to do further configuration to ensure
that clients get the correct DNS, such as:
* sending the clients the VPN DNS server(s) via DHCP, or
* running a local DNS cache (such as powerdns, which can forward
  requests for specific domains to the VPN DNS server(s)

Signed-off-by: Corey Hickey <bugfood-ml@fatooh.org>