allow custom routing table specification via hooks
Using hooks for this is a bit odd--a configuration file would make more
sense, but there currently isn't one. Should I add one?

This is useful on a host that functions as a router for clients that
need VPN access, in which case the router should not actually have
access by default.

In this case, the user will have to do further configuration to ensure
that clients traverse the specified routing table. For example, if
routing table "8" were to be used and client 192.168.0.2 should get a
VPN route:

    # ip rule add from 192.168.0.2 lookup 8

Setting up NAT is probably necessary as well:

    # iptables -t POSTROUTING -o "$TUNDEV" -j MASQUERADE

Signed-off-by: Corey Hickey <bugfood-ml@fatooh.org>