prefer CISCO_SPLIT_DNS to CISCO_DEF_DOMAIN
I do not know if this is the right thing to do. This may need to be
adjusted based on David's advice when I submit it upstream.

This is only implemented for systemd and generic. If this is an
acceptable change, then I can attempt to change the others, but I will
need help testing.

Signed-off-by: Corey Hickey <bugfood-ml@fatooh.org>